email_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
arcadepy
pydantic==2.6.1 